feat: duplicate and remove sticker
