+++ Sources/General/Extensions/ZLEditImageViewController+UIScrollViewDelegate .swift
@@ -1,0 +1,54 @@
+//
+//  ZLEditImageViewController+UIScrollViewDelegate .swift
+//  ZLImageEditor
+//
+//  Created by Rahmat Zulfikri on 10/05/25.
+//
+
+import UIKit
+
+
+// MARK: scroll view delegate
+extension ZLEditImageViewController: UIScrollViewDelegate {
+    public func viewForZooming(in scrollView: UIScrollView) -> UIView? {
+        return containerView
+    }
+    
+    public func scrollViewDidZoom(_ scrollView: UIScrollView) {
+        let offsetX = (scrollView.frame.width > scrollView.contentSize.width) ? (scrollView.frame.width - scrollView.contentSize.width) * 0.5 : 0
+        let offsetY = (scrollView.frame.height > scrollView.contentSize.height) ? (scrollView.frame.height - scrollView.contentSize.height) * 0.5 : 0
+        containerView.center = CGPoint(x: scrollView.contentSize.width * 0.5 + offsetX, y: scrollView.contentSize.height * 0.5 + offsetY)
+    }
+    
+    public func scrollViewDidEndZooming(_ scrollView: UIScrollView, with view: UIView?, atScale scale: CGFloat) {
+        isScrolling = false
+    }
+    
+    public func scrollViewDidScroll(_ scrollView: UIScrollView) {
+        guard scrollView == mainScrollView else {
+            return
+        }
+        isScrolling = true
+    }
+    
+    public func scrollViewDidEndDragging(_ scrollView: UIScrollView, willDecelerate decelerate: Bool) {
+        guard scrollView == mainScrollView else {
+            return
+        }
+        isScrolling = decelerate
+    }
+    
+    public func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
+        guard scrollView == mainScrollView else {
+            return
+        }
+        isScrolling = false
+    }
+    
+    public func scrollViewDidEndScrollingAnimation(_ scrollView: UIScrollView) {
+        guard scrollView == mainScrollView else {
+            return
+        }
+        isScrolling = false
+    }
+}
