+++ Sources/General/ZLArrowView.swift
@@ -50,11 +50,11 @@ class ZLArrowView: ZLBaseStickerView {
              gesScale: state.gesScale,
              gesRotation: state.gesRotation,
              totalTranslationPoint: state.totalTranslationPoint,
-             showBorder: false
+             showBorder: true
         )
         self.backgroundColor = .clear
         self.contentMode = .redraw
-        self.layer.borderWidth = 0 // Ensure no visual border
+        self.layer.borderWidth = 1 // Ensure no visual border
     }
 
     required init?(coder: NSCoder) {
