+++ Sources/General/ZLBaseStickerView.swift
@@ -47,7 +47,7 @@ protocol ZLStickerViewAdditional: NSObject {
     
     func resetState()
     
-    func moveToAshbin()
+    func remove()
     
     func addScale(_ scale: CGFloat)
 }
@@ -73,9 +73,7 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
     var maxGesScale: CGFloat
     
     var originTransform: CGAffineTransform = .identity
-    
-    var timer: Timer?
-    
+        
     var totalTranslationPoint: CGPoint = .zero
     
     var gesTranslationPoint: CGPoint = .zero
@@ -113,10 +111,6 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
     }
     
     weak var delegate: ZLStickerViewDelegate?
-    
-    deinit {
-        cleanTimer()
-    }
     
     class func initWithState(_ state: ZLBaseStickertState) -> ZLBaseStickerView? {
         if let state = state as? ZLTextStickerState {
@@ -158,9 +152,6 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
         
         borderView.layer.borderWidth = borderWidth
         hideBorder()
-        if showBorder {
-            startTimer()
-        }
         
         addGestureRecognizer(tapGes)
         addGestureRecognizer(pinchGes)
@@ -226,9 +217,8 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
     
     @objc func tapAction(_ ges: UITapGestureRecognizer) {
         guard gesIsEnabled else { return }
-        
+        showBorder()
         delegate?.stickerDidTap(self)
-        startTimer()
     }
     
     @objc func pinchAction(_ ges: UIPinchGestureRecognizer) {
@@ -305,12 +295,10 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
     func setOperation(_ isOn: Bool) {
         if isOn, !onOperation {
             onOperation = true
-            cleanTimer()
             borderView.layer.borderColor = UIColor.white.cgColor
             delegate?.stickerBeginOperation(self)
         } else if !isOn, onOperation {
             onOperation = false
-            startTimer()
             delegate?.stickerEndOperation(self, panGes: panGes)
         }
     }
@@ -337,24 +325,15 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
         delegate?.stickerOnOperation(self, panGes: panGes)
     }
     
-    @objc private func hideBorder() {
+    @objc public func showBorder() {
+        borderView.layer.borderColor = UIColor.white.cgColor
+    }
+    
+    @objc public func hideBorder() {
         borderView.layer.borderColor = UIColor.clear.cgColor
     }
     
-    func startTimer() {
-        cleanTimer()
-        borderView.layer.borderColor = UIColor.white.cgColor
-        timer = Timer.scheduledTimer(timeInterval: 2, target: ZLWeakProxy(target: self), selector: #selector(hideBorder), userInfo: nil, repeats: false)
-        RunLoop.current.add(timer!, forMode: .common)
-    }
-    
-    private func cleanTimer() {
-        timer?.invalidate()
-        timer = nil
-    }
-    
     // MARK: UIGestureRecognizerDelegate
-
     func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
         return true
     }
@@ -363,12 +342,10 @@ class ZLBaseStickerView: UIView, UIGestureRecognizerDelegate {
 extension ZLBaseStickerView: ZLStickerViewAdditional {
     func resetState() {
         onOperation = false
-        cleanTimer()
         hideBorder()
     }
     
-    func moveToAshbin() {
-        cleanTimer()
+    func remove() {
         removeFromSuperview()
     }
